+++ packages/ricos-converters/web/src/utils/text-direction.ts
@@ -1,4 +1,5 @@
 import { TextStyle_TextAlignment } from 'ricos-schema';
+import { getTextDirection } from 'wix-rich-content-common';
 
 export const getTextDirectionFromAlignment = (textAlignment: TextStyle_TextAlignment | undefined) =>
   textAlignment === TextStyle_TextAlignment.RIGHT
@@ -6,3 +7,13 @@ export const getTextDirectionFromAlignment = (textAlignment: TextStyle_TextAlign
     : textAlignment === TextStyle_TextAlignment.LEFT
     ? 'ltr'
     : undefined;
+
+export const getListItemDirection = (
+  textAlignment: TextStyle_TextAlignment | undefined,
+  text?: string
+) => {
+  const textDirection = ['ltr', 'rtl'].includes(getTextDirection(text))
+    ? getTextDirection(text)
+    : 'ltr';
+  return getTextDirectionFromAlignment(textAlignment) || textDirection;
+};