+++ CHANGELOG.md
@@ -17,6 +17,11 @@
     Click to see more.
   </summary>
 
+## :bug: Bug Fix
+
+- `tiptap-extensions`
+  - [#4088](https://github.com/wix/ricos/pull/4088) fix headings & lists alignment ,indent & text direction (lists wip)
+
 ## :house: Internal
 
 - `ricos-events`
@@ -38,7 +43,9 @@
 ## :bug: Bug Fix
 
 - `ricos-content`
+
   - [#4103](https://github.com/wix/ricos/pull/4103) fix normalizeUrl func to handle "mailto:" url's
+
 - `ricos-editor`
   - [#4137](https://github.com/wix/ricos/pull/4137) use RicosPortal in TextFormattingToolbars.tsx